+++ CHANGELOG.md
@@ -9,6 +9,8 @@ project adheres to [Semantic Versioning](https://semver.org/).
 ## [0.1.6] - Unreleased
 ### Fixed
 - Action types.
+- Disable action button when it's pressed.
+- UI styles tweaks.
 
 ## [0.1.5] - 2025-06-04
 ### Added